+++ src/data/strikes.ts
