+++ src/endpoints/controls.ts
@@ -1,7 +1,7 @@
 import { RequestHandler } from "express";
-import * as strikeAccess from "../api/strikes";
-import * as meetingAccess from "../api/meetings";
-import * as ogAccess from "../api/ogs";
+import * as strikeAccess from "../data/strikes";
+import * as meetingAccess from "../data/meetings";
+import * as ogAccess from "../data/ogs";
 import { FCMAdmin } from "../services/fcm";
 import { Liveevent } from "../models/liveeventModel";
